+++ ThornClient/Managers/ConfigManager.cs
@@ -3,7 +3,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using BepInEx;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ThornClient.Core;