+++ ThornClient/System/ClickGUIComponents/ConfigurableWindowController.cs
@@ -40,13 +40,16 @@ internal class ConfigurableWindowController : MonoBehaviour {
         if (IsPopup) backBtn.onClick.AddListener(ClickGUI.NavigateBack);
 
         // Body
-        Transform listBody = gameObject.FindRecursive("Scroll View/Viewport/Content/Modules").transform; // Note we reuse ModuleCategory prefab for this
+        Transform listBody =
+            gameObject.FindRecursive("Scroll View/Viewport/Content/Modules")
+                .transform; // Note we reuse ModuleCategory prefab for this
 
         var desc = Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "ModuleDescription"), listBody);
         desc.FindRecursive("DescText").GetComponent<TextMeshProUGUI>().text = TargetConfigurable.Description;
 
         if (TargetConfigurable.HasToggling) {
-            var enabledButton = Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "EnabledButton"), listBody);
+            var enabledButton =
+                Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "EnabledButton"), listBody);
             var enabledButtonComp = enabledButton.AddComponent<EnabledButtonController>();
             enabledButtonComp.Configurable = TargetConfigurable;
         }
@@ -65,23 +68,25 @@ internal class ConfigurableWindowController : MonoBehaviour {
 
     private void Populate(Transform parent, IEnumerable<IConfigurableElement> elements) {
         foreach (var element in elements) {
-            GameObject? wrapper = null;
+            GameObject? wrapper;
+            GameObject? obj = null;
+            if (UICreatorManager.TryGetUICreator(element.GetType(), out var creator) && creator != null) {
+                obj = creator.CreateUI(element);
+            } else {
+                continue;
+            }
 
             if (element is Setting setting) {
+                wrapper = Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "SettingRowWrapper"), parent);                
+                wrapper!.AddComponent<SettingRowController>().TargetSetting = setting;
+                obj?.transform.SetParent(wrapper.transform, false);
+            } else {
                 wrapper = Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "SettingRowWrapper"), parent);
-                wrapper.AddComponent<SettingRowController>().TargetSetting = setting;
-
-                if (ConfigurableElementUICreators.SettingUICreators.TryGetValue(setting.Type, out var createUI)) {
-                    GameObject go = createUI(setting, wrapper.transform);
-                    go.AddComponent<SettingDescriptionController>().TargetSetting = setting;
-                }
-            } else if (element is SettingGroup or ConfigButtonRow or ConfigHeader) {
-                wrapper = Instantiate(AssetManager.Get<GameObject>(ClickGUI.BundleKey, "SettingRowWrapper"), parent);
-                if (ConfigurableElementUICreators.MenuUICreators.TryGetValue(element.GetType(), out var createUI)) {
-                    var go = createUI(element, wrapper.transform);
-                    if (element is not ConfigHeader) go.AddComponent<SettingDescriptionController>().TargetSetting = element;
-                }
+                obj?.transform.SetParent(wrapper!.transform, false);
             }
+
+            if (element is not ConfigHeader && obj != null)
+                obj.AddComponent<SettingDescriptionController>().TargetSetting = element;
 
             if (wrapper != null) {
                 wrapper.UnfuckLayoutHack();